requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-allauth==65.13.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
